Async/Await Support (continuation of #191) (#229)
* it builds!

* remove unused dependencies

* bump dependencies

* reimplement 'exec' endpoint

* update a few more examples

* update remaining examples

* fix doc tests, remove unused 'read' module

* remove feature-gated async closures

* split futures dependency to just 'futures-util'

* update version and readme

* make functions accepting Body generic over Into<Body> again

* update changelog

* reinstate 'unix-socket' feature

* reinstate 'attach' endpoint

* fix clippy lints

* fix documentation typo

* fix container copyfrom/into implementations

* add convenience methods for TtyChunk struct

* remove 'main' from code example to silence clippy lint

* Update hyper to 0.13.1

* Add Send bounds to TtyWriter

* Appease clippy

* Fix examples

* Update issue in changelog

+++ examples/containerdelete.rs
@@ -1,16 +1,14 @@
 use shiplift::Docker;
 use std::env;
-use tokio::prelude::Future;
 
-fn main() {
+#[tokio::main]
+async fn main() {
     let docker = Docker::new();
     let id = env::args()
         .nth(1)
         .expect("You need to specify an container id");
-    let fut = docker
-        .containers()
-        .get(&id)
-        .delete()
-        .map_err(|e| eprintln!("Error: {}", e));
-    tokio::run(fut);
+
+    if let Err(e) = docker.containers().get(&id).delete().await {
+        eprintln!("Error: {}", e)
+    }
 }
